Cleanup: Move array indexing validation out of apis package
We validate that indexing references to array parameters are in bounds,
based on the default values of the parameters and the parameter values
passed in from a PipelineRun/TaskRun. This validation happens in the reconciler,
since it requires comparing the Task/Pipeline spec with the parameters
defined in the TaskRun/PipelineRun.

Prior to this commit, validation code was in the apis package, even though it was called
in the reconciler. This is confusing, because this code cannot be used to validate a Task/Pipeline
spec in isolation; it can only be used in the context of TaskRuns and PipelineRuns.
This commit moves this validation logic to the reconciler and does not introduce any functional changes.
