+++ pkg/apis/pipeline/v1/pipeline_validation_test.go
@@ -3340,183 +3340,6 @@ func getTaskSpec() TaskSpec {
 	}
 }
 
-func TestValidateParamArrayIndex_valid(t *testing.T) {
-	ctx := context.Background()
-	cfg := config.FromContextOrDefaults(ctx)
-	cfg.FeatureFlags.EnableAPIFields = config.BetaAPIFields
-	ctx = config.ToContext(ctx, cfg)
-	for _, tt := range []struct {
-		name     string
-		original PipelineSpec
-		params   Params
-	}{{
-		name: "single parameter",
-		original: PipelineSpec{
-			Params: []ParamSpec{
-				{Name: "first-param", Type: ParamTypeArray, Default: NewStructuredValues("default-value", "default-value-again")},
-				{Name: "second-param", Type: ParamTypeString},
-			},
-			Tasks: []PipelineTask{{
-				Params: Params{
-					{Name: "first-task-first-param", Value: *NewStructuredValues("$(params.first-param[1])")},
-					{Name: "first-task-second-param", Value: *NewStructuredValues("$(params.second-param[0])")},
-					{Name: "first-task-third-param", Value: *NewStructuredValues("static value")},
-				},
-			}},
-		},
-		params: Params{{Name: "second-param", Value: *NewStructuredValues("second-value", "second-value-again")}},
-	}, {
-		name: "single parameter with when expression",
-		original: PipelineSpec{
-			Params: []ParamSpec{
-				{Name: "first-param", Type: ParamTypeArray, Default: NewStructuredValues("default-value", "default-value-again")},
-				{Name: "second-param", Type: ParamTypeString},
-			},
-			Tasks: []PipelineTask{{
-				When: []WhenExpression{{
-					Input:    "$(params.first-param[1])",
-					Operator: selection.In,
-					Values:   []string{"$(params.second-param[0])"},
-				}},
-			}},
-		},
-		params: Params{{Name: "second-param", Value: *NewStructuredValues("second-value", "second-value-again")}},
-	}, {
-		name: "pipeline parameter nested inside task parameter",
-		original: PipelineSpec{
-			Params: []ParamSpec{
-				{Name: "first-param", Type: ParamTypeArray, Default: NewStructuredValues("default-value", "default-value-again")},
-				{Name: "second-param", Type: ParamTypeArray, Default: NewStructuredValues("default-value", "default-value-again")},
-			},
-			Tasks: []PipelineTask{{
-				Params: Params{
-					{Name: "first-task-first-param", Value: *NewStructuredValues("$(input.workspace.$(params.first-param[0]))")},
-					{Name: "first-task-second-param", Value: *NewStructuredValues("$(input.workspace.$(params.second-param[1]))")},
-				},
-			}},
-		},
-		params: nil, // no parameter values.
-	}, {
-		name: "array parameter",
-		original: PipelineSpec{
-			Params: []ParamSpec{
-				{Name: "first-param", Type: ParamTypeArray, Default: NewStructuredValues("default", "array", "value")},
-				{Name: "second-param", Type: ParamTypeArray},
-			},
-			Tasks: []PipelineTask{{
-				Params: Params{
-					{Name: "first-task-first-param", Value: *NewStructuredValues("firstelement", "$(params.first-param)")},
-					{Name: "first-task-second-param", Value: *NewStructuredValues("firstelement", "$(params.second-param[0])")},
-				},
-			}},
-		},
-		params: Params{
-			{Name: "second-param", Value: *NewStructuredValues("second-value", "array")},
-		},
-	}, {
-		name: "parameter evaluation with final tasks",
-		original: PipelineSpec{
-			Params: []ParamSpec{
-				{Name: "first-param", Type: ParamTypeArray, Default: NewStructuredValues("default-value", "default-value-again")},
-				{Name: "second-param", Type: ParamTypeArray},
-			},
-			Finally: []PipelineTask{{
-				Params: Params{
-					{Name: "final-task-first-param", Value: *NewStructuredValues("$(params.first-param[0])")},
-					{Name: "final-task-second-param", Value: *NewStructuredValues("$(params.second-param[1])")},
-				},
-				When: WhenExpressions{{
-					Input:    "$(params.first-param[0])",
-					Operator: selection.In,
-					Values:   []string{"$(params.second-param[1])"},
-				}},
-			}},
-		},
-		params: Params{{Name: "second-param", Value: *NewStructuredValues("second-value", "second-value-again")}},
-	}, {
-		name: "parameter evaluation with both tasks and final tasks",
-		original: PipelineSpec{
-			Params: []ParamSpec{
-				{Name: "first-param", Type: ParamTypeArray, Default: NewStructuredValues("default-value", "default-value-again")},
-				{Name: "second-param", Type: ParamTypeArray},
-			},
-			Tasks: []PipelineTask{{
-				Params: Params{
-					{Name: "final-task-first-param", Value: *NewStructuredValues("$(params.first-param[0])")},
-					{Name: "final-task-second-param", Value: *NewStructuredValues("$(params.second-param[1])")},
-				},
-			}},
-			Finally: []PipelineTask{{
-				Params: Params{
-					{Name: "final-task-first-param", Value: *NewStructuredValues("$(params.first-param[0])")},
-					{Name: "final-task-second-param", Value: *NewStructuredValues("$(params.second-param[1])")},
-				},
-				When: WhenExpressions{{
-					Input:    "$(params.first-param[0])",
-					Operator: selection.In,
-					Values:   []string{"$(params.second-param[1])"},
-				}},
-			}},
-		},
-		params: Params{{Name: "second-param", Value: *NewStructuredValues("second-value", "second-value-again")}},
-	}, {
-		name: "parameter references with bracket notation and special characters",
-		original: PipelineSpec{
-			Params: []ParamSpec{
-				{Name: "first.param", Type: ParamTypeArray, Default: NewStructuredValues("default-value", "default-value-again")},
-				{Name: "second/param", Type: ParamTypeArray},
-				{Name: "third.param", Type: ParamTypeArray, Default: NewStructuredValues("default-value", "default-value-again")},
-				{Name: "fourth/param", Type: ParamTypeArray},
-			},
-			Tasks: []PipelineTask{{
-				Params: Params{
-					{Name: "first-task-first-param", Value: *NewStructuredValues(`$(params["first.param"][0])`)},
-					{Name: "first-task-second-param", Value: *NewStructuredValues(`$(params["second/param"][0])`)},
-					{Name: "first-task-third-param", Value: *NewStructuredValues(`$(params['third.param'][1])`)},
-					{Name: "first-task-fourth-param", Value: *NewStructuredValues(`$(params['fourth/param'][1])`)},
-					{Name: "first-task-fifth-param", Value: *NewStructuredValues("static value")},
-				},
-			}},
-		},
-		params: Params{
-			{Name: "second/param", Value: *NewStructuredValues("second-value", "second-value-again")},
-			{Name: "fourth/param", Value: *NewStructuredValues("fourth-value", "fourth-value-again")},
-		},
-	}, {
-		name: "single parameter in workspace subpath",
-		original: PipelineSpec{
-			Params: []ParamSpec{
-				{Name: "first-param", Type: ParamTypeArray, Default: NewStructuredValues("default-value", "default-value-again")},
-				{Name: "second-param", Type: ParamTypeArray},
-			},
-			Tasks: []PipelineTask{{
-				Params: Params{
-					{Name: "first-task-first-param", Value: *NewStructuredValues("$(params.first-param[0])")},
-					{Name: "first-task-second-param", Value: *NewStructuredValues("static value")},
-				},
-				Workspaces: []WorkspacePipelineTaskBinding{
-					{
-						Name:      "first-workspace",
-						Workspace: "first-workspace",
-						SubPath:   "$(params.second-param[1])",
-					},
-				},
-			}},
-		},
-		params: Params{{Name: "second-param", Value: *NewStructuredValues("second-value", "second-value-again")}},
-	},
-	} {
-		tt := tt // capture range variable
-		t.Run(tt.name, func(t *testing.T) {
-			t.Parallel()
-			err := tt.original.ValidateParamArrayIndex(ctx, tt.params)
-			if err != nil {
-				t.Errorf("ValidateParamArrayIndex() got err %s", err)
-			}
-		})
-	}
-}
-
 func TestPipelineWithBetaFields(t *testing.T) {
 	tts := []struct {
 		name string
@@ -3711,231 +3534,6 @@ func TestPipelineWithBetaFields(t *testing.T) {
 			ctx := config.EnableBetaAPIFields(context.Background())
 			if err := pipeline.Validate(ctx); err != nil {
 				t.Errorf("unexpected error when using beta field: %s", err)
-			}
-		})
-	}
-}
-
-func TestValidateParamArrayIndex_invalid(t *testing.T) {
-	for _, tt := range []struct {
-		name     string
-		original PipelineSpec
-		params   Params
-		expected error
-	}{{
-		name: "single parameter reference out of bound",
-		original: PipelineSpec{
-			Params: []ParamSpec{
-				{Name: "first-param", Type: ParamTypeArray, Default: NewStructuredValues("default-value", "default-value-again")},
-				{Name: "second-param", Type: ParamTypeString},
-			},
-			Tasks: []PipelineTask{{
-				Params: Params{
-					{Name: "first-task-first-param", Value: *NewStructuredValues("$(params.first-param[2])")},
-					{Name: "first-task-second-param", Value: *NewStructuredValues("$(params.second-param[2])")},
-					{Name: "first-task-third-param", Value: *NewStructuredValues("static value")},
-				},
-			}},
-		},
-		params:   Params{{Name: "second-param", Value: *NewStructuredValues("second-value", "second-value-again")}},
-		expected: fmt.Errorf("non-existent param references:[$(params.first-param[2]) $(params.second-param[2])]"),
-	}, {
-		name: "single parameter reference with when expression out of bound",
-		original: PipelineSpec{
-			Params: []ParamSpec{
-				{Name: "first-param", Type: ParamTypeArray, Default: NewStructuredValues("default-value", "default-value-again")},
-				{Name: "second-param", Type: ParamTypeString},
-			},
-			Tasks: []PipelineTask{{
-				When: []WhenExpression{{
-					Input:    "$(params.first-param[2])",
-					Operator: selection.In,
-					Values:   []string{"$(params.second-param[2])"},
-				}},
-			}},
-		},
-		params:   Params{{Name: "second-param", Value: *NewStructuredValues("second-value", "second-value-again")}},
-		expected: fmt.Errorf("non-existent param references:[$(params.first-param[2]) $(params.second-param[2])]"),
-	}, {
-		name: "pipeline parameter reference nested inside task parameter out of bound",
-		original: PipelineSpec{
-			Params: []ParamSpec{
-				{Name: "first-param", Type: ParamTypeArray, Default: NewStructuredValues("default-value", "default-value-again")},
-				{Name: "second-param", Type: ParamTypeArray, Default: NewStructuredValues("default-value", "default-value-again")},
-			},
-			Tasks: []PipelineTask{{
-				Params: Params{
-					{Name: "first-task-first-param", Value: *NewStructuredValues("$(input.workspace.$(params.first-param[2]))")},
-					{Name: "first-task-second-param", Value: *NewStructuredValues("$(input.workspace.$(params.second-param[2]))")},
-				},
-			}},
-		},
-		params:   nil, // no parameter values.
-		expected: fmt.Errorf("non-existent param references:[$(params.first-param[2]) $(params.second-param[2])]"),
-	}, {
-		name: "array parameter reference out of bound",
-		original: PipelineSpec{
-			Params: []ParamSpec{
-				{Name: "first-param", Type: ParamTypeArray, Default: NewStructuredValues("default", "array", "value")},
-				{Name: "second-param", Type: ParamTypeArray},
-			},
-			Tasks: []PipelineTask{{
-				Params: Params{
-					{Name: "first-task-first-param", Value: *NewStructuredValues("firstelement", "$(params.first-param[3])")},
-					{Name: "first-task-second-param", Value: *NewStructuredValues("firstelement", "$(params.second-param[4])")},
-				},
-			}},
-		},
-		params: Params{
-			{Name: "second-param", Value: *NewStructuredValues("second-value", "array")},
-		},
-		expected: fmt.Errorf("non-existent param references:[$(params.first-param[3]) $(params.second-param[4])]"),
-	}, {
-		name: "object parameter reference out of bound",
-		original: PipelineSpec{
-			Params: []ParamSpec{
-				{Name: "first-param", Type: ParamTypeArray, Default: NewStructuredValues("default", "array", "value")},
-				{Name: "second-param", Type: ParamTypeArray},
-			},
-			Tasks: []PipelineTask{{
-				Params: Params{
-					{Name: "first-task-first-param", Value: *NewObject(map[string]string{
-						"val1": "$(params.first-param[4])",
-						"val2": "$(params.second-param[4])",
-					})},
-				},
-			}},
-		},
-		params: Params{
-			{Name: "second-param", Value: *NewStructuredValues("second-value", "array")},
-		},
-		expected: fmt.Errorf("non-existent param references:[$(params.first-param[4]) $(params.second-param[4])]"),
-	}, {
-		name: "parameter evaluation with final tasks reference out of bound",
-		original: PipelineSpec{
-			Params: []ParamSpec{
-				{Name: "first-param", Type: ParamTypeArray, Default: NewStructuredValues("default-value", "default-value-again")},
-				{Name: "second-param", Type: ParamTypeArray},
-			},
-			Finally: []PipelineTask{{
-				Params: Params{
-					{Name: "final-task-first-param", Value: *NewStructuredValues("$(params.first-param[2])")},
-					{Name: "final-task-second-param", Value: *NewStructuredValues("$(params.second-param[2])")},
-				},
-				When: WhenExpressions{{
-					Input:    "$(params.first-param[5])",
-					Operator: selection.In,
-					Values:   []string{"$(params.second-param[1])"},
-				}},
-			}},
-		},
-		params:   Params{{Name: "second-param", Value: *NewStructuredValues("second-value", "second-value-again")}},
-		expected: fmt.Errorf("non-existent param references:[$(params.first-param[2]) $(params.first-param[5]) $(params.second-param[2])]"),
-	}, {
-		name: "parameter evaluation with both tasks and final tasks reference out of bound",
-		original: PipelineSpec{
-			Params: []ParamSpec{
-				{Name: "first-param", Type: ParamTypeArray, Default: NewStructuredValues("default-value", "default-value-again")},
-				{Name: "second-param", Type: ParamTypeArray},
-			},
-			Tasks: []PipelineTask{{
-				Params: Params{
-					{Name: "final-task-first-param", Value: *NewStructuredValues("$(params.first-param[2])")},
-					{Name: "final-task-second-param", Value: *NewStructuredValues("$(params.second-param[2])")},
-				},
-			}},
-			Finally: []PipelineTask{{
-				Params: Params{
-					{Name: "final-task-first-param", Value: *NewStructuredValues("$(params.first-param[3])")},
-					{Name: "final-task-second-param", Value: *NewStructuredValues("$(params.second-param[3])")},
-				},
-				When: WhenExpressions{{
-					Input:    "$(params.first-param[2])",
-					Operator: selection.In,
-					Values:   []string{"$(params.second-param[2])"},
-				}},
-				Matrix: &Matrix{
-					Params: Params{
-						{Name: "final-task-first-param", Value: *NewStructuredValues("$(params.first-param[4])")},
-						{Name: "final-task-second-param", Value: *NewStructuredValues("$(params.second-param[4])")},
-					}},
-			}},
-		},
-		params:   Params{{Name: "second-param", Value: *NewStructuredValues("second-value", "second-value-again")}},
-		expected: fmt.Errorf("non-existent param references:[$(params.first-param[2]) $(params.first-param[3]) $(params.first-param[4]) $(params.second-param[2]) $(params.second-param[3]) $(params.second-param[4])]"),
-	}, {
-		name: "parameter in matrix reference out of bound",
-		original: PipelineSpec{
-			Params: []ParamSpec{
-				{Name: "first-param", Type: ParamTypeArray, Default: NewStructuredValues("default-value", "default-value-again")},
-				{Name: "second-param", Type: ParamTypeArray},
-			},
-			Tasks: []PipelineTask{{
-				Matrix: &Matrix{
-					Params: Params{
-						{Name: "first-task-first-param", Value: *NewStructuredValues("$(params.first-param[2])")},
-						{Name: "first-task-second-param", Value: *NewStructuredValues("static value")},
-					},
-				},
-			}},
-		},
-		params:   Params{{Name: "second-param", Value: *NewStructuredValues("second-value", "second-value-again")}},
-		expected: fmt.Errorf("non-existent param references:[$(params.first-param[2])]"),
-	}, {
-		name: "parameter references with bracket notation and special characters reference out of bound",
-		original: PipelineSpec{
-			Params: []ParamSpec{
-				{Name: "first.param", Type: ParamTypeArray, Default: NewStructuredValues("default-value", "default-value-again")},
-				{Name: "second/param", Type: ParamTypeArray},
-				{Name: "third.param", Type: ParamTypeArray, Default: NewStructuredValues("default-value", "default-value-again")},
-				{Name: "fourth/param", Type: ParamTypeArray},
-			},
-			Tasks: []PipelineTask{{
-				Params: Params{
-					{Name: "first-task-first-param", Value: *NewStructuredValues(`$(params["first.param"][2])`)},
-					{Name: "first-task-second-param", Value: *NewStructuredValues(`$(params["second/param"][2])`)},
-					{Name: "first-task-third-param", Value: *NewStructuredValues(`$(params['third.param'][2])`)},
-					{Name: "first-task-fourth-param", Value: *NewStructuredValues(`$(params['fourth/param'][2])`)},
-					{Name: "first-task-fifth-param", Value: *NewStructuredValues("static value")},
-				},
-			}},
-		},
-		params: Params{
-			{Name: "second/param", Value: *NewStructuredValues("second-value", "second-value-again")},
-			{Name: "fourth/param", Value: *NewStructuredValues("fourth-value", "fourth-value-again")},
-		},
-		expected: fmt.Errorf("non-existent param references:[$(params[\"first.param\"][2]) $(params[\"second/param\"][2]) $(params['fourth/param'][2]) $(params['third.param'][2])]"),
-	}, {
-		name: "single parameter in workspace subpath reference out of bound",
-		original: PipelineSpec{
-			Params: []ParamSpec{
-				{Name: "first-param", Type: ParamTypeArray, Default: NewStructuredValues("default-value", "default-value-again")},
-				{Name: "second-param", Type: ParamTypeArray},
-			},
-			Tasks: []PipelineTask{{
-				Params: Params{
-					{Name: "first-task-first-param", Value: *NewStructuredValues("$(params.first-param[2])")},
-					{Name: "first-task-second-param", Value: *NewStructuredValues("static value")},
-				},
-				Workspaces: []WorkspacePipelineTaskBinding{
-					{
-						Name:      "first-workspace",
-						Workspace: "first-workspace",
-						SubPath:   "$(params.second-param[3])",
-					},
-				},
-			}},
-		},
-		params:   Params{{Name: "second-param", Value: *NewStructuredValues("second-value", "second-value-again")}},
-		expected: fmt.Errorf("non-existent param references:[$(params.first-param[2]) $(params.second-param[3])]"),
-	},
-	} {
-		tt := tt // capture range variable
-		t.Run(tt.name, func(t *testing.T) {
-			t.Parallel()
-			err := tt.original.ValidateParamArrayIndex(context.Background(), tt.params)
-			if d := cmp.Diff(tt.expected.Error(), err.Error()); d != "" {
-				t.Errorf("ValidateParamArrayIndex() errors diff %s", diff.PrintWantGot(d))
 			}
 		})
 	}
